refactor: Enhance user management interface with conditional rendering and improved button labels

+++ app/views/admin/users/index.blade.php
@@ -6,15 +6,17 @@
             <div class="row">
                 <div class="col-12">
                     <div class="card">
-                        <div class="card-body position-relative">
+                        <div class="card-header border-0 position-relative">
                             <h5>
-                                {{__('Users')}}
+                                {{__('Users Management')}}
                             </h5> 
-                            <a class="btn btn-primary position-absolute" 
-                                style="top:0.8rem;right:2rem;" href="/admin/users/add"
-                                data-bs-toggle="modal" data-bs-target="#registerNewUser">
-                                {{__('Add New')}}
-                            </a>              
+                            @if($canCreateUser->status)
+                                <a class="btn btn-primary position-absolute" 
+                                    style="top:0.8rem;right:2rem;" href="/admin/users/add"
+                                    data-bs-toggle="modal" data-bs-target="#registerNewUser">
+                                    {{__('Add New User')}}
+                                </a>
+                            @endif
                         </div>
                     </div>
                 </div>
@@ -34,7 +36,9 @@
                                                 <th>{{__('Email')}}</th>
                                                 <th>{{__('Role')}}</th>
                                                 <th>{{__('Status')}}</th>
-                                                <th>{{__('Actions')}}</th>
+                                                @if($canUpdateUser->status || $canDeleteUser->status)
+                                                    <th></th>
+                                                @endif
                                             </tr>
                                         </thead>
                                         <tbody>
@@ -43,20 +47,25 @@
 
                                                 <tr>
                                                     <td>
-                                                        <a href="/admin/user/{{ $helpers::encode($user->id) }}">
+                                                        <!-- TODO: view user profile -->
+                                                        <!--a href="/admin/user/{{ $helpers::encode($user->id) }}"-->
                                                             {{ $user->fullname }}
-                                                        </a>
+                                                        <!--/a-->
                                                     </td>
                                                     <td> {{ $user->email }} </td>
                                                     <td> {{ $user->user_role->description }} </td>
                                                     <td> {{ $user->status ?? 'Inactive' }} </td>
-                                                    <td>
-                                                        <a href="/admin/user/{{ $helpers::encode($user->id) }}" class="btn btn-primary rounded btn-sm">
-                                                            <i class="fa fa-pencil"></i>
-                                                        </a>
-                                                        <button class="btn btn-danger rounded btn-sm" onclick="deleteUser('{{ $helpers::encode($user->id) }}')">
-                                                            <i class="fa fa-trash"></i>
-                                                        </button>                                                        
+                                                    <td class="text-end">
+                                                        @if($canUpdateUser->status)
+                                                            <a href="/admin/user/{{ $helpers::encode($user->id) }}" class="btn btn-primary rounded btn-sm">
+                                                                <i class="fa fa-pencil"></i>
+                                                            </a>
+                                                        @endif
+                                                        @if($canDeleteUser->status)
+                                                            <button class="btn btn-danger rounded btn-sm" onclick="deleteUser('{{ $helpers::encode($user->id) }}')">
+                                                                <i class="fa fa-trash"></i>
+                                                            </button>
+                                                        @endif                                                        
                                                     </td>
                                                 </tr>
 
@@ -79,57 +88,8 @@
         </div>
     </div>
 
-    <!-- User Registration Modal -->
-    <div class="modal fade" id="registerNewUser" tabindex="-1" aria-labelledby="registerNewUserLabel" aria-hidden="true">
-        <div class="modal-dialog modal-dialog-lg modal-dialog-centered">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="registerNewUserLabel">{{__('Register New User')}}</h5>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                </div>
-                <div class="modal-body">
-
-                    <div class="alert bg-light-warning text-warning" role="alert">
-                        {{__('User registered by admin will automatically be active')}}
-                    </div>
-
-                    <form action="post" name="userRegistrationForm">
-                        @csrf
-                        <div class="row">
-                            <div class="col-md-6 col-sm-12">
-                                <div class="form-group">
-                                    <label class="form-label">Full Name</label>
-                                    <input type="text" name="fullname" class='form-control' placeholder="John Doe" required>
-                                </div>
-                            </div>
-                            <div class="col-md-6 col-sm-12">
-                                <div class="form-group">
-                                    <label class="form-label">Email address</label>
-                                    <input type="text" name="email" class='form-control' placeholder="doe@example.com" required>
-                                </div>
-                            </div>
-                            <div class="col-md-12 col-sm-12">
-                                <div class="form-group">
-                                    <label class="form-label">User role</label>
-                                    <select name="user_role" class="form-select">
-                                        <option value="subscriber" hidden>{{__('Subscriber')}}</option>
-                                        @foreach (\App\Models\Role::all() as $role)
-                                            <option value="{{ $role->name }}">{{ $role->name }}</option>                                                        
-                                        @endforeach
-                                    </select>
-                                </div>
-                            </div>
-                            <div class="col-md-12 col-sm-12">
-                                <div class="form-group">
-                                    <button type="submit" class="btn btn-primary rounded w-100 ">{{__('Create User')}}</button>
-                                </div>
-                            </div>
-                        </div>
-
-                    </form>
-
-                </div>
-            </div>
-        </div>
-    </div>
+    @if($canCreateUser->status)
+        @include('admin.users.partials.add')
+    @endif
+    
 @endsection
